% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{fire data set.}
\format{
A data frame with 15 rows and 2 variables:
\description{
  \item{DISTANCE}{distance from fire station in miles}
  \item{DAMAGE}{X 100 dollars in dama}
  }
  @source description
}
\usage{
fire
}
\description{
A data set about fire
}
\keyword{datasets}
