% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{Title}
\usage{
mybin(iter, n, p)
}
\arguments{
\item{iter}{The number of iterations the function goes through}

\item{n}{the size of the given group}

\item{p}{the probability}
}
\value{
a colored graph
}
\description{
Title
}
