% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myBoot.R
\name{myboot2}
\alias{myboot2}
\title{boot strap method}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, xstat, ...)
}
\arguments{
\item{iter}{number of iterations}

\item{x}{size of the sample}

\item{fun}{what data to use}

\item{alpha}{the cutoff for the data}

\item{...}{}
}
\value{
returns a graph showing a random sample distibuation
}
\description{
boot strap method
}
\examples{
 myboot2(x = ddt$DDT)
}
