% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Project-1 Overbooking problem}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{number of seats in the flight}

\item{gamma}{probabilty the plane will be truly overbooked}

\item{p}{probabilty of a "show"}
}
\value{
two graphs one discrete and one continous and a list of vals
}
\description{
Project-1 Overbooking problem
}
\examples{
ntickets(400, 0.2, 0.95)
}
